requirements.txt
pandas==2.2.2
numpy==2.1.2
scikit-learn==1.5.2
scipy==1.14.1
joblib==1.4.2
matplotlib==3.9.2
plotly==5.24.1
pydeck==0.9.1
fastapi==0.115.0
uvicorn==0.31.1
pydantic==2.9.2
python-dotenv==1.0.1
requests==2.32.3
shap==0.46.0
pyarrow==17.0.0
streamlit==1.39.0
pydeck
requests
